- Improved warning message formatting - Moved quant8 folder to mcu/generic - Modified CBuilder to enable more natural C code generation - Added ArduinoBuilder class to generalize Arduino/C print handling - Reimplemented swap_buffer to use static instead of dynamic memory - Defined dot_product in common.h/c and refactored dependent code - Unified prediction output for C/Arduino examples - Added Doxyfile for documentation - Added license and contributors file - Added SVM support for fixed16 and fixed8 - Implement partial memory statistics for input, output, and parameter data - Add buffer size calculation for the most memory-intensive pipeline layer - Include memory requirements tracking for input/output data across all layers
- Improved warning message formatting
- Moved quant8 folder to mcu/generic
- Modified CBuilder to enable more natural C code generation
- Added ArduinoBuilder class to generalize Arduino/C print handling
- Reimplemented swap_buffer to use static instead of dynamic memory
- Defined dot_product in common.h/c and refactored dependent code
- Unified prediction output for C/Arduino examples
- Added Doxyfile for documentation
- Added license and contributors file
- Added SVM support for fixed16 and fixed8
- Implement partial memory statistics for input, output, and parameter data
- Add buffer size calculation for the most memory-intensive pipeline layer
- Include memory requirements tracking for input/output data across all layers

+++ export examples/outputs/Prj-Conv2D_Net_IMG2D/common.h
@@ -37,7 +37,7 @@ typedef struct{
 
 void prepare_buffers();
 
-void * swap_alloc(size_t s);
+void * swap_alloc(uint32_t s);
 
 
 /*
